separated functionalities and renamed variables
